+++ bot/general/values.py
@@ -7,34 +7,6 @@ from semantic_version import Version
 
 BOT_VERSION = Version("1.0.0")
 
-
-help_message = """Here are the list of commands available to you:
-
-General Commands
-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-
-/start \-\- Starts the bot up\.
-/about \-\- Tells you a bit about the bot\.
-/help \-\- The bot will send this message that you're reading right now\.
-/quip \-\- The bot will hit you up with a mighty fine quip or just an okay one\.
-/f \-\- The bot will reply with a gif to help pay respect\.
-/mf \-\- The bot will reply with a sad\.\.\. sad\.\. voice note\.
-/age \-\- The bot will reply with how old the bot is\.
-
-Notify Group Commands \- Send a message while tagging preset users \(Meant to be used in group chats\)
-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-
-/add\_notify\_group \-\- create a notify group for that group chat
-/modify\_notify\_group \-\- modify any of your notify groups for a specific group chat
-`/invite\_to\_notify\_group example\_group\_name @example\_username1 @example\_username2` \-\- Invites the tagged users to your notify group
-/list\_notify\_groups \-\- Lists all the notify groups in a group chat
-`/notify example\_group\_name optional\_message` \-\- Notifies all members within a notify group with an optional message
-
-Status Commands \- Get status of players on Xbox Live and PSN\.
-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-
-/add\_status\_user \-\- Create a status user belonging to a group chat or private chat
-/modify\_status\_user \-\- Modify a status user belonging to a group chat or private chat
-/status \-\- Lists the status of all status users within a group chat or private chat
-"""
-
 disses = [
     "I'mmmmmmmmmmm sooorrryyyyyyyyy for your loss",
     "It looks like your face caught on fire and someone tried to put it out with a fork",
@@ -75,12 +47,21 @@ BELL_EMOJI = u"\U0001F514"
 ENVELOPE_EMOJI = u"\U00002709"
 INFORMATION_DESK_PERSON_EMOJI = u"\U0001F481"
 NOTEPAD_EMOJI = u"\U0001F5D2"
+LEDGER_EMOJI = u"\U0001F4D2"
+WINK_EMOJI = u"\U0001F609"
+ZANY_EMOJI = u"\U0001F92A"
+MONEY_BAG_EMOJI = u"\U0001F4B0"
+QUESTION_MARK_EMOJI = u"\U00002753"
+
 
 ADMIN_LIST = [247340182]
 
 TOKEN = os.environ.get("GUB_BOT_TOKEN")
+PAYMENT_TOKEN = os.environ.get("GUB_PAYMENT_TOKEN")
+FEEDBACK_TOKEN = os.environ.get("GUB_FEEDBACK_TOKEN")
 XBOX_CLIENT_SECRET_EXPIRY_DATE = datetime.datetime.strptime("Jun 16, 2024", "%b %d, %Y").date()
 
 OBIWAN_HELLO_THERE_GIF_FILEPATH = "media/obiwans-hello-there.mp4"
 F_TO_PAY_RESPECT_GIF_FILEPATH = "media/f-to-pay-respect.gif"
 MISSION_FAILED_AUDIO_FILEPATH = "media/mission-failed.mp3"
+WHAT_DID_HE_SAY_VIDEO_FILEPATH = "media/what-did-he-say.mp4"