+++ infrastructure/main.tf
@@ -157,6 +157,8 @@ resource "aws_ecs_task_definition" "gsm-task-definition" {
       "essential" : true,
       "environment" : [
         { "name" : "GUB_BOT_TOKEN", "value" : var.GUB_TG_BOT_TOKEN },
+        { "name" : "GUB_PAYMENT_TOKEN", "value" : var.GUB_TG_PAYMENT_TOKEN },
+        { "name" : "GUB_FEEDBACK_TOKEN", "value" : var.GUB_TG_FEEDBACK_TOKEN },
         { "name" : "GSM_DB_URL_WITHOUT_USERNAME_AND_PASSWORD", "value" : "${mongodbatlas_serverless_instance.gsm-db.connection_strings_standard_srv}" },
         { "name" : "GSM_DB_USERNAME", "value" : var.MONGODBATLAS_DB_USER_USERNAME },
         { "name" : "GSM_DB_PASSWORD", "value" : "${random_password.mongodbatlas_password.result}" },